+++ src/pymovements/__init__.py
@@ -39,6 +39,7 @@ from pymovements.events import EventGazeProcessor
 from pymovements.events import EventProcessor
 from pymovements.gaze import Experiment
 from pymovements.gaze import EyeTracker
+from pymovements.gaze import Gaze
 from pymovements.gaze import GazeDataFrame
 from pymovements.gaze import Screen
 from pymovements.measure import register_sample_measure
@@ -63,6 +64,7 @@ __all__ = [
     'Experiment',
     'EyeTracker',
     'Screen',
+    'Gaze',
     'GazeDataFrame',
 
     'exceptions',
