+++ src/pymovements/gaze/__init__.py
@@ -28,6 +28,7 @@
     Experiment
     EyeTracker
     Screen
+    Gaze
     GazeDataFrame
 
 .. rubric:: Transformations
@@ -79,6 +80,7 @@ from pymovements.gaze import transforms
 from pymovements.gaze import transforms_numpy
 from pymovements.gaze.experiment import Experiment
 from pymovements.gaze.eyetracker import EyeTracker
+from pymovements.gaze.gaze import Gaze
 from pymovements.gaze.gaze_dataframe import GazeDataFrame
 from pymovements.gaze.integration import from_numpy
 from pymovements.gaze.integration import from_pandas
@@ -93,6 +95,7 @@ __all__ = [
     'EyeTracker',
     'from_numpy',
     'from_pandas',
+    'Gaze',
     'GazeDataFrame',
     'Screen',
     'transforms_numpy',
