+++ src/pymovements/gaze/integration.py
@@ -17,7 +17,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-"""Module to create a GazeDataFrame from a numpy array."""
+"""Module to create a Gaze from a numpy array."""
 from __future__ import annotations
 
 from typing import Literal
@@ -29,7 +29,7 @@ import polars as pl
 from pymovements._utils import _checks
 from pymovements.events.frame import EventDataFrame
 from pymovements.gaze.experiment import Experiment
-from pymovements.gaze.gaze_dataframe import GazeDataFrame
+from pymovements.gaze.gaze import Gaze
 
 
 def from_numpy(
@@ -54,8 +54,8 @@ def from_numpy(
         velocity_columns: list[str] | None = None,
         acceleration_columns: list[str] | None = None,
         distance_column: str | None = None,
-) -> GazeDataFrame:
-    """Get a :py:class:`~pymovements.gaze.GazeDataFrame` from a numpy array.
+) -> Gaze:
+    """Get a :py:class:`~pymovements.gaze.Gaze` from a numpy array.
 
     There are two mutually exclusive ways of conversion.
 
@@ -121,7 +121,7 @@ def from_numpy(
 
     Returns
     -------
-    GazeDataFrame
+    Gaze
         Returns gaze data frame read from numpy array.
 
     Examples
@@ -236,7 +236,7 @@ def from_numpy(
 
     if data is not None:
         df = pl.from_numpy(data=data, schema=schema, orient=orient)
-        return GazeDataFrame(
+        return Gaze(
             data=df,
             experiment=experiment,
             events=events,
@@ -297,7 +297,7 @@ def from_numpy(
         distance_column = 'distance'
 
     df = pl.concat(dfs, how='horizontal')
-    return GazeDataFrame(
+    return Gaze(
         data=df,
         experiment=experiment,
         events=events,
@@ -325,8 +325,8 @@ def from_pandas(
         velocity_columns: list[str] | None = None,
         acceleration_columns: list[str] | None = None,
         distance_column: str | None = None,
-) -> GazeDataFrame:
-    """Get a :py:class:`~pymovements.gaze.GazeDataFrame` from a pandas DataFrame.
+) -> Gaze:
+    """Get a :py:class:`~pymovements.gaze.Gaze` from a pandas DataFrame.
 
     Parameters
     ----------
@@ -364,11 +364,11 @@ def from_pandas(
 
     Returns
     -------
-    GazeDataFrame
+    Gaze
         Returns gaze data frame read from pandas data frame.
     """
     df = pl.from_pandas(data=data)
-    return GazeDataFrame(
+    return Gaze(
         data=df,
         experiment=experiment,
         events=events,
