+++ tests/unit/events/event_aoi_mapping_test.py
@@ -17,7 +17,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-"""Test all GazeDataFrame functionality."""
+"""Test all Gaze functionality."""
 import polars as pl
 import pytest
 from polars.testing import assert_frame_equal
@@ -740,14 +740,14 @@ def test_map_to_aois_raises_value_error():
         height_column='height',
         page_column='page',
     )
-    gaze_df = pm.gaze.io.from_csv(
+    gaze = pm.gaze.io.from_csv(
         'tests/files/judo1000_example.csv',
         **{'separator': '\t'},
         position_columns=['x_left', 'y_left', 'x_right', 'y_right'],
     )
 
     with pytest.raises(ValueError) as excinfo:
-        gaze_df.map_to_aois(aoi_df, eye='right', gaze_type='')
+        gaze.map_to_aois(aoi_df, eye='right', gaze_type='')
     msg, = excinfo.value.args
     assert msg == 'neither position nor pixel in gaze dataframe, one needed for mapping'
