+++ tests/unit/gaze/apply_test.py
@@ -17,7 +17,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-"""Test GazeDataFrame detect method."""
+"""Test Gaze detect method."""
 import numpy as np
 import polars as pl
 import pytest
@@ -115,7 +115,7 @@ from pymovements.synthetic import step_function
         pytest.param(
             'downsample',
             {'factor': 2},
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     {
                         'time': np.arange(1000, 1010, 1),
@@ -125,7 +125,7 @@ from pymovements.synthetic import step_function
                 ),
                 pixel_columns=['x_pix', 'y_pix'],
             ),
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     {
                         'time': np.arange(1000, 1010, 2),
@@ -141,7 +141,7 @@ from pymovements.synthetic import step_function
         pytest.param(
             'pix2deg',
             {},
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     {
                         'time': [1000, 1000],
@@ -152,7 +152,7 @@ from pymovements.synthetic import step_function
                 experiment=pm.Experiment(100, 100, 100, 100, 100, 'center', 1000),
                 pixel_columns=['x_pix', 'y_pix'],
             ),
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     {
                         'time': [1000, 1000],
@@ -171,7 +171,7 @@ from pymovements.synthetic import step_function
         pytest.param(
             'deg2pix',
             {'pixel_origin': 'center'},
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     {
                         'time': [1000, 1000],
@@ -182,7 +182,7 @@ from pymovements.synthetic import step_function
                 experiment=pm.Experiment(100, 100, 100, 100, 100, 'center', 1000),
                 position_columns=['x_dva', 'y_dva'],
             ),
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     {
                         'time': [1000, 1000],
@@ -201,7 +201,7 @@ from pymovements.synthetic import step_function
         pytest.param(
             'pos2vel',
             {'method': 'preceding'},
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     {
                         'trial_id': [1, 1, 1, 2, 2, 2],
@@ -214,7 +214,7 @@ from pymovements.synthetic import step_function
                 trial_columns='trial_id',
                 position_columns=['x', 'y'],
             ),
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     {
                         'trial_id': [1, 1, 1, 2, 2, 2],
