+++ tests/unit/gaze/gaze_unnest_test.py
@@ -17,7 +17,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-"""Test GazeDataFrame.unnest()."""
+"""Test Gaze.unnest()."""
 import polars as pl
 import pytest
 from polars.testing import assert_frame_equal
@@ -229,8 +229,8 @@ import pymovements as pm
 
     ],
 )
-def test_gaze_dataframe_unnest_has_expected_frame(init_data, unnest_kwargs, expected):
-    gaze = pm.GazeDataFrame(init_data)
+def test_gaze_unnest_has_expected_frame(init_data, unnest_kwargs, expected):
+    gaze = pm.Gaze(init_data)
     gaze.unnest(**unnest_kwargs)
     assert_frame_equal(gaze.frame, expected)
 
@@ -351,10 +351,10 @@ def test_gaze_dataframe_unnest_has_expected_frame(init_data, unnest_kwargs, expe
         ),
     ],
 )
-def test_gaze_dataframe_unnest_has_expected_frame_multiple_unnest(
+def test_gaze_unnest_has_expected_frame_multiple_unnest(
         init_data, unnest_kwargs, expected,
 ):
-    gaze = pm.GazeDataFrame(init_data)
+    gaze = pm.Gaze(init_data)
     gaze.unnest(**unnest_kwargs)
     assert_frame_equal(gaze.frame, expected)
 
@@ -443,9 +443,9 @@ def test_gaze_dataframe_unnest_has_expected_frame_multiple_unnest(
     ],
 
 )
-def test_gaze_dataframe_unnest_errors(init_data, unnest_kwargs, exception, exception_msg):
+def test_gaze_unnest_errors(init_data, unnest_kwargs, exception, exception_msg):
     with pytest.raises(exception) as exc_info:
-        gaze = pm.GazeDataFrame(init_data)
+        gaze = pm.Gaze(init_data)
         gaze.unnest(**unnest_kwargs)
 
     msg, = exc_info.value.args
@@ -466,11 +466,11 @@ def test_gaze_dataframe_unnest_errors(init_data, unnest_kwargs, exception, excep
     ],
 
 )
-def test_gaze_dataframe_unnest_invalid_number_of_components(
+def test_gaze_unnest_invalid_number_of_components(
         init_data, unnest_kwargs, n_components, exception, exception_msg,
 ):
     with pytest.raises(exception) as exc_info:
-        gaze = pm.GazeDataFrame(init_data)
+        gaze = pm.Gaze(init_data)
         gaze.n_components = n_components
         gaze.unnest(**unnest_kwargs)
