+++ tests/unit/gaze/measure_test.py
@@ -17,7 +17,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-"""Test GazeDataFrame measure_samples method."""
+"""Test Gaze measure_samples method."""
 import numpy as np
 import polars as pl
 import pytest
@@ -34,7 +34,7 @@ def my_test_measure(column: str) -> pl.Expr:
     ('gaze', 'method', 'kwargs', 'expected'),
     [
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': [1000, 1001, 1002, 1003]}, schema={'A': pl.Int64}),
             ),
             'null_ratio',
@@ -44,7 +44,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': [1000, None, None, 1003]}, schema={'A': pl.Int64}),
             ),
             'null_ratio',
@@ -54,7 +54,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': [1.0, np.nan, np.nan, 1.3]}, schema={'A': pl.Float64}),
             ),
             'null_ratio',
@@ -64,7 +64,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': [None, None, None, None]}, schema={'A': pl.Int64}),
             ),
             'null_ratio',
@@ -74,7 +74,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': [0.0, 0.1, 0.2, 0.3]}, schema={'A': pl.Float64}),
             ),
             'null_ratio',
@@ -84,7 +84,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': [0.0, None, None, 0.3]}, schema={'A': pl.Float64}),
             ),
             'null_ratio',
@@ -94,7 +94,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': [None, None, None, None]}, schema={'A': pl.Float64}),
             ),
             'null_ratio',
@@ -104,7 +104,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': ['a', 'b', 'c', 'd']}, schema={'A': pl.Utf8}),
             ),
             'null_ratio',
@@ -114,7 +114,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': ['a', None, None, 'd']}, schema={'A': pl.Utf8}),
             ),
             'null_ratio',
@@ -124,7 +124,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': [None, None, None, None]}, schema={'A': pl.Utf8}),
             ),
             'null_ratio',
@@ -134,7 +134,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     data={'t': [1000, 1001, 1002], 'x': [0.1, 0.2, 0.3], 'y': [0.1, 0.2, 0.3]},
                 ),
@@ -148,7 +148,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     data={
                         't': [1000, 1001, 1002], 'x': [None, None, None], 'y': [None, None, None],
@@ -164,7 +164,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     data={
                         't': [1000, 1001], 'x': [0.1, None], 'y': [0.2, None],
@@ -180,7 +180,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.DataFrame(
                     data={
                         't': [1000, 1001], 'x': [0.1, np.nan], 'y': [0.2, np.nan],
@@ -196,7 +196,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     data={
                         't': [1000, 1001], 'x': [0.1, None], 'y': [0.2, None],
@@ -212,7 +212,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     data={
                         't': [1000, 1001], 'x': [0.1, None], 'y': [0.2, None],
@@ -228,7 +228,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     data={'A': [1000, 1001, 1002, 1003], 'trial': [1, 1, 1, 1]},
                     schema={'A': pl.Int64, 'trial': pl.Int64},
@@ -242,7 +242,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(
                     data={'A': [1000, 1001, None, None], 'trial': [1, 1, 2, 2]},
                     schema={'A': pl.Int64, 'trial': pl.Int64},
@@ -256,7 +256,7 @@ def my_test_measure(column: str) -> pl.Expr:
         ),
 
         pytest.param(
-            pm.GazeDataFrame(
+            pm.Gaze(
                 data=pl.from_dict(data={'A': [1000, 1001, 1002, 1003]}, schema={'A': pl.Int64}),
             ),
             my_test_measure,
