+++ tests/unit/plotting/heatmap_test.py
@@ -28,7 +28,7 @@ import pytest
 from matplotlib import figure
 
 from pymovements.gaze import Experiment
-from pymovements.gaze.gaze_dataframe import GazeDataFrame
+from pymovements.gaze.gaze import Gaze
 from pymovements.plotting import heatmap
 
 
@@ -56,15 +56,15 @@ def args_fixture(experiment_fixture, request):
         },
     )
 
-    # Init a GazeDataFrame
-    gdf = GazeDataFrame(
+    # Init a Gaze
+    gaze = Gaze(
         data=df,
         experiment=experiment_fixture,
         pixel_columns=pixel_columns,
         position_columns=position_columns,
     )
 
-    return gdf, request.param
+    return gaze, request.param
 
 
 @pytest.mark.parametrize(
@@ -208,7 +208,7 @@ def test_heatmap_no_experiment_property():
         },
     )
 
-    gdf = GazeDataFrame(data=df, pixel_columns=['x_pix', 'y_pix'], experiment=None)
+    gaze = Gaze(data=df, pixel_columns=['x_pix', 'y_pix'], experiment=None)
 
     with pytest.raises(ValueError):
-        heatmap(gdf, show=False)
+        heatmap(gaze, show=False)
